+++ capacitor.config.ts
@@ -1,11 +1,11 @@
 import type { CapacitorConfig } from '@capacitor/cli'
 
 const config: CapacitorConfig = {
-  appId: 'com.fm619tech.paperphoneplus',
-  appName: 'PaperPhonePlus',
+  appId: 'com.fm619tech.paperphonelite',
+  appName: 'PaperPhoneLite',
   webDir: 'dist',
   server: {
-    // HTTPS scheme is required for WebRTC getUserMedia() and crypto.subtle
+    // HTTPS scheme is required for WebRTC getUserMedia() and crypto.subtle.
     iosScheme: 'https',
   },
   plugins: {
@@ -15,9 +15,6 @@ const config: CapacitorConfig = {
       splashFullScreen: true,
       splashImmersive: true,
       backgroundColor: '#ffffff',
-    },
-    PushNotifications: {
-      presentationOptions: ['badge', 'sound', 'alert'],
     },
   },
   ios: {